Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
boto3==1.35.99
# Additional dependencies for image processing and model training
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
